// osoma/src/domain/dvc/dvc-proxy.generated.ts
export interface components {
    schemas: {
// ...
        Metric: {
            id: string;
            description: string;
            /** @enum {string} */
            type: "METRIC" | "ENV_PARAM";
        };
        MetricsResponse: components["schemas"]["Metric"][];
        RfidInterval: {
            rfid: string;
            start: number;
            stop: number;
        };
        GlobResult: {
            uuid: string;
            humanReadableId: string;
            /** @enum {string} */
            status: "RUNNING" | "OUT_OF_RACK" | "TERMINATED";
            /** Format: date-time */
            lastEventDate: string;
            /** Format: date-time */
            registrationEventDate: string;
            rfids: components["schemas"]["RfidInterval"][];
        };
        GlobResults: components["schemas"]["GlobResult"][];
        GlobSearchRequest: {
            patterns: string[];
        };
        CageSearchResponse: {
            cages: components["schemas"]["GlobResults"];
            /** Format: date-time */
            min_date: string | null;
            /** Format: date-time */
            max_date: string | null;
        };
        TaskSubmission: {
            /** @enum {string} */
            aggrType: "HOUR" | "MINUTE";
            ids: string[];
            metrics: string[];
            outputFilename: string;
            /** @enum {string} */
            resourceType: "CAGE" | "ANIMAL";
            start: string;
            stop: string;
            /** @enum {string} */
            partitionType: "SINGLE_FILE" | "SINGLE_FILE_BY_DAY" | "SINGLE_FILE_BY_CAGE";
        };
        TaskSubmissionResponse: {
            taskId: number;
            requestId: string;
        };
        TaskState: {
            id: number;
            /** @enum {string} */
            state: "SUBMITTED" | "RUNNING" | "COMPLETED" | "FAILED";
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            finishedAt?: string;
            errorMessage?: string;
            fileSize?: number;
            checksum?: string;
        };
        TaskStateResponse: components["schemas"]["TaskState"] | components["schemas"]["TaskState"][];
        TestApiKeyObject: {
            [key: string]: unknown;
        };
        TestApiKeyResponse: string | components["schemas"]["TestApiKeyObject"];
        ErrorResponse: {
            message: string;
            detail: string;
        };
    };
// ...
}